Add auth.go and operations.go files
Added the auth.go file to handle authentication on the server in the future. This will ensure that the server can only be accessed by authorised people. There will also be different access levels for resources.
Also added the operations.go file to handle requests to the server involving modification of a table.