Add MakeTableNew() function
Added the MakeTableNew() function in operations.go to create a new Table from column and row data. It is used to respond with a list of tables or dbs as a table.
Having a gotables.conf file is now optional. If no gotables.conf file is present the default config will be used and a warning will be displayed when starting the server.
Added the option in the config to set the log path. Logs will be stored in this location.