Add TableJSON struct
Added the TableJSON struct that is used to convert a table to json and vice versa. This allows the fields in the Table struct to be private which ensures that the fields cannot be directly edited from outside. There are now new methods for the Table struct to get and set the fields as well as two functions (Ttoj() and Jtot()) to convert between Table and TableJSON.