Fix index out of range error
There was an error when running getDBs() or getTables() in the operations/gt_get package. This has been resolved by appending the entire row to the row variable instead of appending to rows[i] (which did not exist as row was initialized with length 0).
The sendJson() function in the server package (now renamed to sendTable()) did not work correctly. This has benn resolved by specifying that the data parameter has to be of type fs.Table.