Change datatype of fs.Column.Type
Previously the Type field of the Column struct was just any value of a datatype. This has now been changed to store the actual datatype (reflect.Type). Additionally, "interface{}" has been replaced with "any" everywhere in the project as this increases readability for non-go developers as it might be more clear what "any" means compared to "interface{}".