Convert datatype in gt_post
Previously, all attempts to change the value of a non-string field or create a new row with non-string column Types failed, as all values extracted from the query string were perceived to be strings. THere is now a new convert() function in gt_post to try to convert the values to their actual datatype.