Create new datatype "num"
There is now a new option for column.Type. It is called "num" and was created to merge the "int" and "flt" datatypes into one. There was a bug when decoding a json file that integers were treated as float64. As json is used to store tables and json is unable to differentiate between integers and floats, this decision has been made. The "int" and "flt" datatypes are still detected but will both just use float64 as the new "num" datatype does.